# Mutable/Immutable refactor and GetImmutable snapshots