This fixes #88 and #68, with a side effect of resolving #50 and #87.