@jlandrews 2018-08-01
So I've run the "fullbench" benchmarks (which took more than 32gb of ram, so we might want to trim them down a bit), and it seems like performance is quite similar to develop.

That doesn't account for actual tree size differences though, and I think I was wrong to add the nodeDB reference to each node, since it's a waste of memory to have it on each one. I only did that because I thought it was nicer to have node.getLeftNode() and node.getRightNode() without arguments of the tree. That's not a necessary change though, so I've reverted the particular commit that did it.