@alexanderbez 2018-08-01
Don't have too much context here, but the changes and discussion look very clean and concise. Does any relevant documentation need to be updated (i.e. readme, docs, etc...)? Also, @jlandrews, do you think the benchmarks should also be run with `--benchmem` and compare against the base branch to see improvements?