@jlandrews 2018-08-01
Reran with -benchmem and compared results. Overall it looks pretty similar, but there are some metrics that changed by 10-20% either direction. Not sure how to interpret it though, since it seems like these tests are non-deterministic and so there may be some variance anyway...

Also I will check for docs to be updated. doc.go definitely needs to be updated and I can do that now, but the biggest piece of documentation is the spec which is still in the pipeline.

```
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkRandomBytes/random-4-24                                          72.0          74.8          +3.89%
BenchmarkRandomBytes/random-16-24                                         118           121           +2.54%
BenchmarkRandomBytes/random-32-24                                         156           166           +6.41%
BenchmarkRandomBytes/random-100-24                                        370           376           +1.62%
BenchmarkRandomBytes/random-1000-24                                       2958          2943          -0.51%
BenchmarkSmall/memdb-1000-100-4-10/query-miss-24                          4420          4081          -7.67%
BenchmarkSmall/memdb-1000-100-4-10/query-hits-24                          5272          5112          -3.03%
BenchmarkSmall/memdb-1000-100-4-10/update-24                              159192        157089        -1.32%
BenchmarkSmall/memdb-1000-100-4-10/block-24                               24824001      25654741      +3.35%
BenchmarkSmall/goleveldb-1000-100-4-10/query-miss-24                      6368          7127          +11.92%
BenchmarkSmall/goleveldb-1000-100-4-10/query-hits-24                      8741          8713          -0.32%
BenchmarkSmall/goleveldb-1000-100-4-10/update-24                          109113        103509        -5.14%
BenchmarkSmall/goleveldb-1000-100-4-10/block-24                           17755475      17394312      -2.03%
BenchmarkSmall/leveldb-1000-100-4-10/query-miss-24                        6584          6214          -5.62%
BenchmarkSmall/leveldb-1000-100-4-10/query-hits-24                        7898          8254          +4.51%
BenchmarkSmall/leveldb-1000-100-4-10/update-24                            102305        107058        +4.65%
BenchmarkSmall/leveldb-1000-100-4-10/block-24                             16187510      17031744      +5.22%
BenchmarkMedium/memdb-100000-100-16-40/query-miss-24                      9595          10064         +4.89%
BenchmarkMedium/memdb-100000-100-16-40/query-hits-24                      10834         11143         +2.85%
BenchmarkMedium/memdb-100000-100-16-40/update-24                          1261394       1303374       +3.33%
BenchmarkMedium/memdb-100000-100-16-40/block-24                           198937291     190258294     -4.36%
BenchmarkMedium/goleveldb-100000-100-16-40/query-miss-24                  22673         22452         -0.97%
BenchmarkMedium/goleveldb-100000-100-16-40/query-hits-24                  28135         28301         +0.59%
BenchmarkMedium/goleveldb-100000-100-16-40/update-24                      294852        296013        +0.39%
BenchmarkMedium/goleveldb-100000-100-16-40/block-24                       34807659      35855483      +3.01%
BenchmarkMedium/leveldb-100000-100-16-40/query-miss-24                    22686         22802         +0.51%
BenchmarkMedium/leveldb-100000-100-16-40/query-hits-24                    27668         31757         +14.78%
BenchmarkMedium/leveldb-100000-100-16-40/update-24                        287108        297615        +3.66%
BenchmarkMedium/leveldb-100000-100-16-40/block-24                         35596044      36791150      +3.36%
BenchmarkLarge/memdb-1000000-100-16-40/query-miss-24                      15911         15781         -0.82%
BenchmarkLarge/memdb-1000000-100-16-40/query-hits-24                      15748         15750         +0.01%
BenchmarkLarge/memdb-1000000-100-16-40/update-24                          5161409       5256647       +1.85%
BenchmarkLarge/memdb-1000000-100-16-40/block-24                           514923704     534785952     +3.86%
BenchmarkLarge/goleveldb-1000000-100-16-40/query-miss-24                  60866         59997         -1.43%
BenchmarkLarge/goleveldb-1000000-100-16-40/query-hits-24                  49123         51637         +5.12%
BenchmarkLarge/goleveldb-1000000-100-16-40/update-24                      478900        476117        -0.58%
BenchmarkLarge/goleveldb-1000000-100-16-40/block-24                       53327657      56340657      +5.65%
BenchmarkLarge/leveldb-1000000-100-16-40/query-miss-24                    57290         62528         +9.14%
BenchmarkLarge/leveldb-1000000-100-16-40/query-hits-24                    48691         50966         +4.67%
BenchmarkLarge/leveldb-1000000-100-16-40/update-24                        381530        456299        +19.60%
BenchmarkLarge/leveldb-1000000-100-16-40/block-24                         56220875      58929008      +4.82%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-miss-24        23263         20237         -13.01%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-hits-24        25201         26440         +4.92%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/update-24            283769        276849        -2.44%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/block-24             33455540      33371134      -0.25%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/query-miss-24       23683         23127         -2.35%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/query-hits-24       32174         30518         -5.15%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/update-24           317281        317968        +0.22%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/block-24            44231625      40372117      -8.73%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/query-miss-24      35564         34792         -2.17%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/query-hits-24      45093         45317         +0.50%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/update-24          574978        521571        -9.29%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/block-24           76693083      78074233      +1.80%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/query-miss-24     147724        160833        +8.87%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/query-hits-24     169279        172494        +1.90%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/update-24         3882880       3285298       -15.39%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/block-24          520453137     482757364     -7.24%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/query-miss-24         22097         22448         +1.59%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/query-hits-24         27734         32738         +18.04%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/update-24             511317        515236        +0.77%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/block-24              2681891       2702588       +0.77%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/query-miss-24        23397         23648         +1.07%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/query-hits-24        28137         28103         -0.12%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/update-24            349234        351358        +0.61%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/block-24             9405218       9918965       +5.46%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/query-miss-24       22684         24165         +6.53%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/query-hits-24       28836         29367         +1.84%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/update-24           279935        291198        +4.02%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/block-24            34924252      35941442      +2.91%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/query-miss-24       21974         22196         +1.01%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/query-hits-24       29079         28725         -1.22%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/update-24           200319        200024        -0.15%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/block-24            143250792     152324790     +6.33%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/query-miss-24      22070         22555         +2.20%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/query-hits-24      28305         28427         +0.43%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/update-24          150990        143810        -4.76%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/block-24           416917935     408154340     -2.10%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-miss-24        19841         20109         +1.35%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-hits-24        25320         25620         +1.18%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/update-24            310418        276745        -10.85%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/block-24             33174907      33831408      +1.98%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/query-miss-24       23820         23057         -3.20%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/query-hits-24       28973         28850         -0.42%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/update-24           315650        310652        -1.58%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/block-24            41392373      40865126      -1.27%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/query-miss-24      33961         33848         -0.33%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/query-hits-24      45861         43692         -4.73%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/update-24          594813        588636        -1.04%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/block-24           77458855      82922740      +7.05%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/query-miss-24     148305        147862        -0.30%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/query-hits-24     169529        178420        +5.24%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/update-24         3540659       3606703       +1.87%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/block-24          515252342     459105626     -10.90%

benchmark                                                                 old allocs     new allocs     delta
BenchmarkRandomBytes/random-4-24                                          1              1              +0.00%
BenchmarkRandomBytes/random-16-24                                         1              1              +0.00%
BenchmarkRandomBytes/random-32-24                                         1              1              +0.00%
BenchmarkRandomBytes/random-100-24                                        1              1              +0.00%
BenchmarkRandomBytes/random-1000-24                                       1              1              +0.00%
BenchmarkSmall/memdb-1000-100-4-10/query-miss-24                          9              9              +0.00%
BenchmarkSmall/memdb-1000-100-4-10/query-hits-24                          12             12             +0.00%
BenchmarkSmall/memdb-1000-100-4-10/update-24                              764            763            -0.13%
BenchmarkSmall/memdb-1000-100-4-10/block-24                               120912         120940         +0.02%
BenchmarkSmall/goleveldb-1000-100-4-10/query-miss-24                      13             13             +0.00%
BenchmarkSmall/goleveldb-1000-100-4-10/query-hits-24                      18             18             +0.00%
BenchmarkSmall/goleveldb-1000-100-4-10/update-24                          254            253            -0.39%
BenchmarkSmall/goleveldb-1000-100-4-10/block-24                           39158          39338          +0.46%
BenchmarkSmall/leveldb-1000-100-4-10/query-miss-24                        14             13             -7.14%
BenchmarkSmall/leveldb-1000-100-4-10/query-hits-24                        18             18             +0.00%
BenchmarkSmall/leveldb-1000-100-4-10/update-24                            254            254            +0.00%
BenchmarkSmall/leveldb-1000-100-4-10/block-24                             39097          39144          +0.12%
BenchmarkMedium/memdb-100000-100-16-40/query-miss-24                      10             10             +0.00%
BenchmarkMedium/memdb-100000-100-16-40/query-hits-24                      12             12             +0.00%
BenchmarkMedium/memdb-100000-100-16-40/update-24                          4746           4746           +0.00%
BenchmarkMedium/memdb-100000-100-16-40/block-24                           795945         795943         -0.00%
BenchmarkMedium/goleveldb-100000-100-16-40/query-miss-24                  27             26             -3.70%
BenchmarkMedium/goleveldb-100000-100-16-40/query-hits-24                  35             35             +0.00%
BenchmarkMedium/goleveldb-100000-100-16-40/update-24                      592            594            +0.34%
BenchmarkMedium/goleveldb-100000-100-16-40/block-24                       67622          67658          +0.05%
BenchmarkMedium/leveldb-100000-100-16-40/query-miss-24                    25             27             +8.00%
BenchmarkMedium/leveldb-100000-100-16-40/query-hits-24                    35             35             +0.00%
BenchmarkMedium/leveldb-100000-100-16-40/update-24                        593            594            +0.17%
BenchmarkMedium/leveldb-100000-100-16-40/block-24                         67118          67963          +1.26%
BenchmarkLarge/memdb-1000000-100-16-40/query-miss-24                      20             20             +0.00%
BenchmarkLarge/memdb-1000000-100-16-40/query-hits-24                      15             15             +0.00%
BenchmarkLarge/memdb-1000000-100-16-40/update-24                          20570          20570          +0.00%
BenchmarkLarge/memdb-1000000-100-16-40/block-24                           2069282        2069277        -0.00%
BenchmarkLarge/goleveldb-1000000-100-16-40/query-miss-24                  82             82             +0.00%
BenchmarkLarge/goleveldb-1000000-100-16-40/query-hits-24                  62             62             +0.00%
BenchmarkLarge/goleveldb-1000000-100-16-40/update-24                      836            838            +0.24%
BenchmarkLarge/goleveldb-1000000-100-16-40/block-24                       98883          95529          -3.39%
BenchmarkLarge/leveldb-1000000-100-16-40/query-miss-24                    76             81             +6.58%
BenchmarkLarge/leveldb-1000000-100-16-40/query-hits-24                    60             61             +1.67%
BenchmarkLarge/leveldb-1000000-100-16-40/update-24                        736            841            +14.27%
BenchmarkLarge/leveldb-1000000-100-16-40/block-24                         101281         100887         -0.39%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-miss-24        27             26             -3.70%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-hits-24        34             34             +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/update-24            563            565            +0.36%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/block-24             64612          65264          +1.01%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/query-miss-24       28             28             +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/query-hits-24       37             37             +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/update-24           616            616            +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-1000/block-24            73233          73717          +0.66%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/query-miss-24      30             30             +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/query-hits-24      39             39             +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/update-24          752            695            -7.58%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-10000/block-24           92942          93850          +0.98%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/query-miss-24     64             65             +1.56%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/query-hits-24     70             71             +1.43%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/update-24         639            621            -2.82%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100000/block-24          102744         84213          -18.04%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/query-miss-24         27             25             -7.41%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/query-hits-24         35             35             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/update-24             842            843            +0.12%
BenchmarkLevelDBBatchSizes/goleveldb-100000-5-16-40/block-24              4350           4385           +0.80%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/query-miss-24        26             27             +3.85%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/query-hits-24        35             35             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/update-24            704            703            -0.14%
BenchmarkLevelDBBatchSizes/goleveldb-100000-25-16-40/block-24             18575          18689          +0.61%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/query-miss-24       27             27             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/query-hits-24       35             35             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/update-24           595            594            -0.17%
BenchmarkLevelDBBatchSizes/goleveldb-100000-100-16-40/block-24            67455          67049          -0.60%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/query-miss-24       26             26             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/query-hits-24       36             35             -2.78%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/update-24           454            448            -1.32%
BenchmarkLevelDBBatchSizes/goleveldb-100000-400-16-40/block-24            273493         273242         -0.09%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/query-miss-24      26             26             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/query-hits-24      35             35             +0.00%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/update-24          326            320            -1.84%
BenchmarkLevelDBBatchSizes/goleveldb-100000-2000-16-40/block-24           824207         815927         -1.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-miss-24        24             26             +8.33%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/query-hits-24        34             34             +0.00%
BenchmarkLevelDBLargeData/goleveldb-50000-100-32-100/update-24            564            566            +0.35%
…